requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
python-multipart==0.0.6

# ML/Data processing - Updated versions
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
openai-whisper
spacy>=3.7.0
numpy>=1.24.0
pandas>=2.0.0

# NLP & Text processing
nltk>=3.8.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.0

# Vector database
faiss-cpu>=1.7.0

# API clients
yt-dlp
requests>=2.31.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Optional
langchain>=0.0.300
langchain-community>=0.0.10